+++ src/main/java/de/jeff_media/angelchest/DebugFormatter.java
@@ -1,0 +1,50 @@
+package de.jeff_media.angelchest;
+
+import java.io.PrintWriter;
+import java.io.StringWriter;
+import java.text.MessageFormat;
+import java.util.Date;
+import java.util.logging.Formatter;
+import java.util.logging.LogRecord;
+
+public class DebugFormatter extends Formatter {
+
+    private final Date date = new Date();
+    private static final String format = "{0,time}";
+    private MessageFormat formatter;
+    private final Object args[] = new Object[1];
+    private final String newLine = System.lineSeparator();
+
+    @Override
+    public synchronized String format(LogRecord record) {
+        StringBuilder sb = new StringBuilder();
+        date.setTime(record.getMillis());
+        args[0] = date;
+
+        // Date and time
+        StringBuffer text = new StringBuffer("[");
+        if(formatter == null) {
+            formatter = new MessageFormat(format);
+        }
+        formatter.format(args,text,null);
+        sb.append(text).append("] ");
+
+
+
+        String message = formatMessage(record);
+
+        sb.append(message);
+        sb.append(newLine);
+        if (record.getThrown() != null) {
+            try {
+                StringWriter sw = new StringWriter();
+                PrintWriter pw = new PrintWriter(sw);
+                record.getThrown().printStackTrace(pw);
+                pw.close();
+                sb.append(sw);
+            } catch (Exception ex) {
+            }
+        }
+        return sb.toString();
+    }
+}